deposit: static and async typeahead fields
