+++ cap/static/js/cap/controllers/capFormAutofillCtrl.js
@@ -27,122 +27,15 @@
 ///////////////////////////////////////////
 // CAP app Form Autofill Controller
 
-var capFormAutofillCtrl = function( $scope, $interpolate , $http){
-    $scope.getData = function(ref, val, model, form) {
-        var _params = {};
-        _params[ref.paramKey] = val;
-        return $http.get(ref.url, {
-            params: _params
-        }).then(function(response){
-            return response.data.map(function(item){
-                if (ref["displayKey"]){
-                    var _item = item;
-                    angular.forEach(ref["displayKey"], function(key){
-                        _item = _item[key];
-                    });
-                    return _item
-                }
-                else{
-                    return item;
-                }
-            });
-        });
-    };
-
-
-    // Needs to be specified with the following options to function
-    //
-    // "type": "cap:formAutofill",
-    // "isAsync": "true",
-    // "ref": {
-    //   "url": "//maps.googleapis.com/maps/api/geocode/json",
-    //   "paramKey": "address",
-    //   "displayKey": ["formatted_address"],
-    //   "map": [
-    //     {
-    //       "s": "place_id",
-    //       "t": ["basic_info", "abstract"]
-    //     },
-    //     {
-    //       "s": ["geometry", "location_type"],
-    //       "t": ["#", "conclusion"]
-    //     },
-    //     {
-    //       "s": ["geometry", "location", "lat"],
-    //       "t": ["#", "software", "name"]
-    //     }
-    //   ]
-    // }
-    $scope.onSelect = function(_item, _model, _label, model, form) {
-        var ref = form.ref;
-        var _params = {};
-        _params[ref.paramKey] = _item;
-
-        $http.get(ref.url, {
-            params: _params
-        }).then(function(response){
-            if( ref && ref.map){
-                angular.forEach(ref.map, function(mapping){
-                    var model_ref = model;
-                    var resp = {};
-
-                    if (mapping["s"]){
-                        if (angular.isArray(mapping["s"])){
-                            resp = response.data;
-                            angular.forEach(mapping["s"], function(key){
-                                resp = resp[key];
-                            });
-                        }
-                        else{
-                            // [TOFIX] for now gets only first result
-                            if (response.data)
-                                resp = response.data[mapping["s"]];
-                        }
-                    }
-
-                    // format response if array or object
-                    if (angular.isArray(resp)){
-                        resp = resp.join("; ");
-                    }
-                    else if (angular.isObject(resp)){
-                        resp = "";
-                    }
-
-                    if (mapping["t"]){
-                        if (mapping["t"][0] === "#"){
-                            _mapping = form.key.slice(0,-1);
-                            _mapping = _mapping.concat((mapping["t"]).slice(1));
-                        }
-                        else{
-                            _mapping = mapping["t"];
-                        }
-
-                        var m_length = _mapping.length;
-                        angular.forEach(_mapping, function(key, index){
-                            if (model_ref[key]){
-                                model_ref = model_ref[key];
-                            }
-                            else if(index+1 !== m_length){
-                                model_ref = model_ref[key] = {};
-                            }
-                            else if(index+1 === m_length){
-                                model_ref = model_ref[key] = resp;
-                                $scope.$emit('fieldAutofilled', resp, {
-                                    "key": _mapping
-                                });
-                            }
-                        });
-                    }
-                });
-            };
-        });
-    };
+var capFormAutofillCtrl = function( $scope, $interpolate , $http, capAutofill){
+    $scope.onSelect = capAutofill.autofill;
 };
 
 capFormAutofillCtrl.$inject = [
     '$scope',
     '$interpolate',
-    '$http'
+    '$http',
+    'capAutofill'
 ];
 
 angular.module('cap.controllers')
