+++ cap/static/templates/cap_records_js/decorators/formAutofill.html
@@ -16,8 +16,7 @@
              type="{{form.type}}"
              step="any"
              sf-changed="form"
-             uib-typeahead="item for item in getData(form.ref, $viewValue , model, form)"
-             typeahead-on-select="onSelect($item, $model, $label, model, form)"
+             ng-change="onSelect($$value$$, $model, $label, model, form)"
              placeholder="{{form.placeholder}}"
              class="form-control {{form.fieldHtmlClass}}"
              id="{{form.key.slice(-1)[0]}}"
@@ -28,7 +27,6 @@
              name="{{form.key.slice(-1)[0]}}"
              autocomplete="false"
              aria-describedby="{{form.key.slice(-1)[0] + 'Status'}}">
-            <button type="button" ng-click="onSelect($$value$$, $model, $label, model, form)" class="btn autofill-btn">Autofill</button>
       <span ng-if="form.feedback !== false" class="form-control-feedback" ng-class="evalInScope(form.feedback) || {'glyphicon': true, 'glyphicon-ok': hasSuccess(), 'glyphicon-remove': hasError() }" aria-hidden="true">
       </span>
       <span ng-if="hasError() || hasSuccess()" id="{{form.key.slice(-1)[0] + 'Status'}}" class="sr-only">
