+++ cap/static/templates/cap_records_js/decorators/typeahead.html
@@ -2,33 +2,33 @@
      ng-controller="typeaheadCtrl"
      ng-class="{'has-error': form.disableErrorState !== true && hasError(), 'has-success': form.disableSuccessState !== true && hasSuccess(), 'has-feedback': form.feedback !== false }" >
      <div class="col-sm-12 cap-label">
-       <label class="control-label {{ form.labelHtmlClass }}"
-              ng-class="{'sr-only': !showTitle(), 'field-required': form.required}"
-              for="{{form.key.slice(-1)[0]}}">
-           <span ng-if="!form.notitle">{{form.title}}</span>
-       </label>
-       <div ng-if="form.description" ><i class="fa fa-minus"></i><div class="help-block" show-more><div sf-message="form.description"></div></div></div >
+         <label class="control-label {{ form.labelHtmlClass }}"
+                ng-class="{'sr-only': !showTitle(), 'field-required': form.required}"
+                for="{{form.key.slice(-1)[0]}}">
+             <span ng-if="!form.notitle">{{form.title}}</span>
+         </label>
+         <div ng-if="form.description" ><i class="fa fa-minus"></i><div class="help-block" show-more><div sf-message="form.description"></div></div></div >
      </div>
-    <div class="col-sm-12 {{ form.fieldWrapperHtmlClass }}">
-        <input ng-if="!form.fieldAddonLeft && !form.fieldAddonRight"
-               ng-show="form.key"
-               type="{{form.type}}"
-               sf-changed="form"
-               uib-typeahead="item for item in typeaheadData | filter:$viewValue"
-               typeahead-on-select="onSelect($item, $model, $label, model, form)"
-               placeholder="{{form.placeholder}}"
-               class="form-control {{form.fieldHtmlClass}}"
-               id="{{form.key.slice(-1)[0]}}"
-               ng-model-options="form.ngModelOptions"
-               ng-model="$$value$$"
-               ng-disabled="form.readonly"
-               schema-validate="form"
-               name="{{form.key.slice(-1)[0]}}"
-               aria-describedby="{{form.key.slice(-1)[0] + 'Status'}}">
-        <span ng-if="form.feedback !== false" class="form-control-feedback" ng-class="evalInScope(form.feedback) || {'glyphicon': true, 'glyphicon-ok': hasSuccess(), 'glyphicon-remove': hasError() }" aria-hidden="true">
-        </span>
-        <span ng-if="hasError() || hasSuccess()" id="{{form.key.slice(-1)[0] + 'Status'}}" class="sr-only">
-            {{ hasSuccess() ? '(success)' : '(error)' }}
-        </span>
-    </div>
+     <div class="col-sm-12 {{ form.fieldWrapperHtmlClass }}">
+         <input ng-if="!form.fieldAddonLeft && !form.fieldAddonRight"
+                ng-show="form.key"
+                type="{{form.type}}"
+                sf-changed="form"
+                uib-typeahead="item for item in getData($viewValue)"
+                typeahead-on-select="onSelect($item, $model, $label, model, form)"
+                placeholder="{{form.placeholder}}"
+                class="form-control {{form.fieldHtmlClass}}"
+                id="{{form.key.slice(-1)[0]}}"
+                ng-model-options="form.ngModelOptions"
+                ng-model="$$value$$"
+                ng-disabled="form.readonly"
+                schema-validate="form"
+                name="{{form.key.slice(-1)[0]}}"
+                aria-describedby="{{form.key.slice(-1)[0] + 'Status'}}">
+         <span ng-if="form.feedback !== false" class="form-control-feedback" ng-class="evalInScope(form.feedback) || {'glyphicon': true, 'glyphicon-ok': hasSuccess(), 'glyphicon-remove': hasError() }" aria-hidden="true">
+         </span>
+         <span ng-if="hasError() || hasSuccess()" id="{{form.key.slice(-1)[0] + 'Status'}}" class="sr-only">
+             {{ hasSuccess() ? '(success)' : '(error)' }}
+         </span>
+     </div>
 </div>
